uefi: Move various types to the `uefi::boot` module
The types are publicly re-exported from `uefi::table::boot`, so there should be
no breaking change to the public API.
